new_version_with_notes/practice/python/requirements.txt
numpy==1.24.2
Pillow==7.0.0
